refactoring training / eval scripts
